Preserve minion cache flag

+++ salt/key.py
@@ -483,10 +483,15 @@ class Key(object):
                                         pubkey,
                                         sig_path)
 
-    def check_minion_cache(self):
+    def check_minion_cache(self, preserve_minions=None):
         '''
         Check the minion cache to make sure that old minion data is cleared
+
+        Optionally, pass in a list of minions which should have their caches
+        preserved. To preserve all caches, set __opts__['preserve_minion_cache']
         '''
+        if preserve_minions is None:
+            preserve_minions = []
         m_cache = os.path.join(self.opts['cachedir'], self.ACC)
         if not os.path.isdir(m_cache):
             return
@@ -494,9 +499,9 @@ class Key(object):
         minions = []
         for key, val in keys.items():
             minions.extend(val)
-        if self.opts.get('preserve_minion_cache', False):
+        if not self.opts.get('preserve_minion_cache', False) or not preserve_minion_cache:
             for minion in os.listdir(m_cache):
-                if minion not in minions:
+                if minion not in minions and minion not in preserve_minions:
                     shutil.rmtree(os.path.join(m_cache, minion))
 
     def check_master(self):
@@ -722,10 +727,12 @@ class Key(object):
                 pass
         return self.list_keys()
 
-    def delete_key(self, match=None, match_dict=None):
+    def delete_key(self, match=None, match_dict=None, preserve_minions=False):
         '''
         Delete public keys. If "match" is passed, it is evaluated as a glob.
         Pre-gathered matches can also be passed via "match_dict".
+
+        To preserve the master caches of minions who are matched, set preserve_minions
         '''
         if match is not None:
             matches = self.name_match(match)
@@ -743,7 +750,7 @@ class Key(object):
                     self.event.fire_event(eload, tagify(prefix='key'))
                 except (OSError, IOError):
                     pass
-        self.check_minion_cache()
+        self.check_minion_cache(preserve_minions=matches.get('minions', []))
         if self.opts.get('rotate_aes_key'):
             salt.crypt.dropfile(self.opts['cachedir'], self.opts['user'], self.opts['sock_dir'])
         return (
@@ -894,7 +901,7 @@ class RaetKey(Key):
         rejected = os.path.join(self.opts['pki_dir'], self.REJ)
         return accepted, pre, rejected
 
-    def check_minion_cache(self):
+    def check_minion_cache(self, preserve_minion_cache=False):
         '''
         Check the minion cache to make sure that old minion data is cleared
         '''
